LPS-123794 Adds custom `useDrag` and `useDrop` and creates more item.types
Created a generic useDrop/useDrag hook for dispatching different actions depending of the type of the Drag.

You may notice We can have two types of Drag:

* You can drag between fields(useDrag from EditorVariant.Column and Pages)

* You can drag between a sidebar and a builder(useDrag defined on Sidebar components)

This knowledge is necessary for deciding which drag action I need to call. The main idea is to call the action ended with `MOVE` when dropping something initialized with values from the drag between fields and `ADD` when dropping something initialized with values from drag from the sidebar.

Also, We can have three different types of a drag item:

* fieldType:
When the item to be dragged is a field/sidebar field

* dataDefinitionField:
When the item to be dragged came from App Builder CustomObjectList sidebar

In the future it will be replaced by just `fieldset` or `fieldtype` depending on the context

* fieldset:
When the item to be dragged came from a fieldset.

Currently, there are 7 types of Drag, combining the already defined types of field/action:

`DragTypes.DRAG_FIELD_TYPE_ADD`:
 `DragTypes.DRAG_DATA_DEFINITION_FIELD_MOVE`:

`DragTypes.DRAG_FIELD_TYPE_MOVE`:

`DragTypes.DRAG_FIELDSET_MOVE`:

`DragTypes.DRAG_DATA_DEFINITION_FIELD_ADD`:

`DragTypes.DRAG_FIELD_TYPE_ADD`:

`DragTypes.DRAG_FIELDSET_ADD`: