LPS-123793 Adds the FormFieldSettings component
This component maintains a FormProvider to handle the structure of the Field's settingsContext, every Field contains the structure of a form that is used within the Sidebar to render a form where it allows editing the properties of the field rendered in the Builder.

FieldSettings needs its own Provider and reducers, it manipulates settingsContext page data and has different behaviors. But he also needs to talk to the root FormProvider and sync some data.

The use of ConfigProvider is still due to the fact that the Sidebar is in Metal and is not rendered on the same structure and cannot have visibility of the application's root Config.

Ideally, we need to have a more elegant way of keeping two FormProvider taking care of two different data but that are dependents.

Some reasons for having the sync is that the Builder root FormProvider reducers have the responsibility to update the information within the settingsContext which could be avoided and done within the FieldSettings reducers to avoid creating any state sync components. This is a research aria of the structural problem that needs to be studied and that could not be done during the migration from Metal.js to React