LPS-123793 Creates the Adapter component to use the Sidebar in Metal.js in React
This is a temporary Adapter that allows you to render the current Forms Sidebar in Metal.js in the React structure. Once we have the Data Engine Sidebar we can remove that implementation.

Some changes to the Sidebar in Metal.js were also made to replace the event names, also removing the Sidebar from Main.es.js that will be rendered in a React component. Present in the next commits.