LPS-124147 Adds the implementation of ToastProvider to handle publication notifications for a form
The old implementation used an old ClayAlert v2 and used the wrong Lexicon pattern to display the success message for publishing the form, so applying the Lexicon pattern we use Toast with an action so that the user can open the form.

A ToastProvider was implemented instead of using the infra openToast because it has not yet implemented Toast with actions. We can remove this Provider when it is implemented.