LPS-129879 Adding some properties to the schemas
Adding some properties that App Builder needs and probably the Data Engine as well, all of these properties respect the structure of the Data Engine.

The implementation of `dataDefinition.dataDefinitionFields` is not yet complete, it is necessary to merge the fields that are inside` pages` which are the most updated fields, and also to make the structure conversion.