LPS-129879 Implements the conversion and merge of fields with dataDefinitionFields
This operation will happen only once and the next calls are from the cache, the value will be revalidated by Schema that makes a comparison by reference of the Schema's props with the state, any changes in these properties the Schema is recreated.

This implementation uses the Data Converter method to convert a DDM field to dataDefinition and at the same time merge fields that are not inside pages, this is for use cases like App Builder.

The Data Converter should only be called from here in the near future and then removed together with Schema as components and reducers migrate to the Data Engine's data structure.