LPS-130483 Removes old and unnecessary calls to the compatibility layer in DataLayoutBuilderInstanceProvider
There are some calls that are no longer needed as before due to the implementation of Metal.js, so we should take it out and it doesn't make sense to keep it or simulate it in the compatibility layer because it won't have any effect just more line and call overhead.

There are still two `useEffect` that need to be changed but for now we don't have an equivalent yet, it will be added in future interactions.

The implementation to change `fieldActions` will be recreated to allow building a fieldAction according to the active field and avoiding things like App Builder is doing here.